LayoutTests/imported/w3c/web-platform-tests/referrer-policy/strict-origin/attr-referrer/same-origin/http-http/img-tag/no-redirect/insecure-protocol.http-expected.txt



FAIL The referrer URL is origin when a
                                 document served over http requires an http
                                 sub-resource via img-tag using the attr-referrer
                                 delivery method with no-redirect and when
                                 the target request is same-origin. assert_equals: Reported Referrer URL is 'origin'. expected "http://localhost:8800/" but got "http://localhost:8800/referrer-policy/strict-origin/attr-referrer/same-origin/http-http/img-tag/no-redirect/insecure-protocol.http.html"
FAIL The referrer URL is origin when a
                                 document served over http requires an http
                                 sub-resource via img-tag using the attr-referrer
                                 delivery method with no-redirect and when
                                 the target request is same-origin. (srcdoc iframe inherits parent) assert_equals: Reported Referrer URL is 'origin'. expected "http://localhost:8800/" but got "http://localhost:8800/referrer-policy/strict-origin/attr-referrer/same-origin/http-http/img-tag/no-redirect/insecure-protocol.http.html"
PASS The referrer URL is origin when a
                                 document served over http requires an http
                                 sub-resource via img-tag using the attr-referrer
                                 delivery method with no-redirect and when
                                 the target request is same-origin. (overridden by srcdoc iframe)
